requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
psycopg[binary]
pytest
alembic
google-api-python-client
google-auth
PyJWT